LayoutTests/imported/w3c/web-platform-tests/offscreen-canvas/text/2d.text.align.invalid.worker-expected.txt


FAIL 2d assert_equals: ctx.textAlign === 'start' (got bogus[string], expected start[string]) expected "start" but got "bogus"
